[test] Extract Papaya test data
