+++ packages/sirius-web/backend/sirius-web/src/test/java/org/eclipse/sirius/web/application/controllers/deck/PapayaDeckControllerIntegrationTests.java
@@ -75,7 +75,7 @@ public class PapayaDeckControllerIntegrationTests extends AbstractIntegrationTes
 
     @Test
     @DisplayName("Given a deck representation, when we subscribe to its event, then the representation data are received")
-    @Sql(scripts = {"/scripts/initialize.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(scripts = {"/scripts/papaya.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts = {"/scripts/cleanup.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD, config = @SqlConfig(transactionMode = SqlConfig.TransactionMode.ISOLATED))
     public void givenDeckRepresentationWhenWeSubscribeToItsEventThenTheRepresentationDataAreReceived() {
         var flux = this.givenSubscriptionToDeck();
