+++ packages/sirius-web/backend/sirius-web/src/test/java/org/eclipse/sirius/web/application/controllers/diagrams/ExpandCollapseDiagramControllerTests.java
@@ -90,7 +90,7 @@ public class ExpandCollapseDiagramControllerTests extends AbstractIntegrationTes
 
     @Test
     @DisplayName("Given a diagram with collapsed nodes by default, when it is opened, then some nodes are collapsed")
-    @Sql(scripts = {"/scripts/initialize.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(scripts = {"/scripts/papaya.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts = {"/scripts/cleanup.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD, config = @SqlConfig(transactionMode = SqlConfig.TransactionMode.ISOLATED))
     public void givenDiagramWithCollapsedNodesByDefaultWhenItIsOpenedThenSomeNodesAreCollapsed() {
         var flux = this.givenSubscriptionToExpandedCollapseDiagram();
@@ -117,7 +117,7 @@ public class ExpandCollapseDiagramControllerTests extends AbstractIntegrationTes
 
     @Test
     @DisplayName("Given a diagram with collapsed nodes by default, when a tool expanding nodes is invoked, then collapsed nodes are expanded")
-    @Sql(scripts = {"/scripts/initialize.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(scripts = {"/scripts/papaya.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts = {"/scripts/cleanup.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD, config = @SqlConfig(transactionMode = SqlConfig.TransactionMode.ISOLATED))
     public void givenDiagramWithCollapsedNodesByDefaultWhenToolExpandingNodesIsInvokedThenCollapsedNodesAreExpanded() {
         var flux = this.givenSubscriptionToExpandedCollapseDiagram();
@@ -161,7 +161,7 @@ public class ExpandCollapseDiagramControllerTests extends AbstractIntegrationTes
 
     @Test
     @DisplayName("Given a diagram with collapsed nodes by default, when a tool collapsing nodes is invoked, then expanded nodes are collapsed")
-    @Sql(scripts = {"/scripts/initialize.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @Sql(scripts = {"/scripts/papaya.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts = {"/scripts/cleanup.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD, config = @SqlConfig(transactionMode = SqlConfig.TransactionMode.ISOLATED))
     public void givenDiagramWithCollapsedNodesByDefaultWhenToolCollapsingNodesIsInvokedThenExpandedNodesAreCollapsed() {
         var flux = this.givenSubscriptionToExpandedCollapseDiagram();
